// src/app/home/enums/atstatus.ts
export enum ATStatus {
  // eslint-disable-next-line @typescript-eslint/naming-convention
  OK = 'OK',
  // eslint-disable-next-line @typescript-eslint/naming-convention
  SENDING = 'SENDING',
  // eslint-disable-next-line @typescript-eslint/naming-convention
  SENDED = 'SENDED',
  // eslint-disable-next-line @typescript-eslint/naming-convention
  CPU_BUSY = 'CPU_BUSY',
  // eslint-disable-next-line @typescript-eslint/naming-convention
  RF_BUSY = 'RF_BUSY',
  // eslint-disable-next-line @typescript-eslint/naming-convention
  SYMBLE = 'SYMBLE',
  // eslint-disable-next-line @typescript-eslint/naming-convention
  PARA = 'PARA',
  // eslint-disable-next-line @typescript-eslint/naming-convention
  CMD = 'CMD'
}